nvim: rework config based on lazy-nvim
This is a starting point to make a nvim setup that is similar to my Doom
Emacs setup. We are taking on a lot here, but I really want to give
neovim another shot based on that it looks to have improved a lot with
lazy-nvim and my Emacs setup is a bit buggy and slow sometimes.